requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
sphinx-material
sphinx-argparse

nltk==3.5
rouge_score==0.0.4
datasets==1.0.2
torch==1.6.0
transformers==3.2.0
pytorch_lightning==0.9.0
